Comment out the terrain to make the engine work
Focus on the Deffered rendering
