+++ Core/src/Assets/AssetHandle.cpp
@@ -1,0 +1,89 @@
+//
+// Created by Sayama on 06/01/2025.
+//
+
+#include "Voxymore/Assets/AssetHandle.hpp"
+#include "Voxymore/Core/Logger.hpp"
+#include "Voxymore/Core/Macros.hpp"
+
+namespace Voxymore::Core {
+
+    std::string AssetTypeToString(AssetType type)
+	{
+		switch (type) {
+			case AssetType::None: return "AssetType::None";
+			case AssetType::Scene: return "AssetType::Scene";
+			case AssetType::Texture2D: return "AssetType::Texture2D";
+			case AssetType::CubeMap: return "AssetType::CubeMap";
+			case AssetType::Material: return "AssetType::Material";
+			case AssetType::Shader: return "AssetType::Shader";
+			case AssetType::ComputeShader: return "AssetType::ComputeShader";
+			case AssetType::ShaderSource: return "AssetType::ShaderSource";
+			case AssetType::Model: return "AssetType::Model";
+			case AssetType::Mesh: return "AssetType::Mesh";
+		}
+		VXM_CORE_ASSERT(false, "The type '{}' is unknown", (uint16_t)type);
+		return "AssetType::<Unknown>";
+	}
+
+	bool TryAssetTypeFromString(const std::string& assetType, AssetType& type)
+	{
+		if(assetType == "AssetType::None") {
+			type = AssetType::None;
+			return true;
+		}
+		else if(assetType == "AssetType::Scene") {
+			type = AssetType::Scene;
+			return true;
+		}
+		else if(assetType == "AssetType::Texture2D") {
+			type = AssetType::Texture2D;
+			return true;
+		}
+		else if(assetType == "AssetType::CubeMap") {
+			type = AssetType::CubeMap;
+			return true;
+		}
+		else if(assetType == "AssetType::Material") {
+			type = AssetType::Material;
+			return true;
+		}
+		else if(assetType == "AssetType::Shader") {
+			type = AssetType::Shader;
+			return true;
+		}
+		else if(assetType == "AssetType::ComputeShader") {
+			type = AssetType::ComputeShader;
+			return true;
+		}
+		else if(assetType == "AssetType::ShaderSource") {
+			type = AssetType::ShaderSource;
+			return true;
+		}
+		else if(assetType == "AssetType::Model") {
+			type = AssetType::Model;
+			return true;
+		}
+		else if(assetType == "AssetType::Mesh") {
+			type = AssetType::Mesh;
+			return true;
+		}
+		return false;
+	}
+
+	AssetType AssetTypeFromString(const std::string& assetType)
+	{
+		AssetType t = AssetType::None;
+		if(!TryAssetTypeFromString(assetType, t)) {
+			VXM_CORE_ERROR("The type '{}' is unknown", assetType);
+			return AssetType::None;
+		}
+		return t;
+	}
+
+
+    std::string AssetTypeToPayloadID(AssetType assetType)
+    {
+        return std::string("VXM_ASSET_") + std::to_string(static_cast<uint16_t>(assetType));
+    }
+}
