+++ Core/src/Scene/Scene.cpp
@@ -381,13 +381,13 @@ namespace Voxymore::Core
 				Renderer::Submit(pc.GetMesh(), transform.GetTransform(), static_cast<int>(entity));
 			}
 		}
-
+#ifdef VXM_TERRAIN
 		if (m_Terrain && m_Terrain->IsValid()) {
 			m_Terrain->Bind();
 			m_Terrain->Draw();
 			m_Terrain->Unbind();
 		}
-
+#endif
 		// Draw Physics
 		if(JPH::DebugRenderer::sInstance == nullptr) {
 			JPH::DebugRenderer::sInstance = new PhysicsDebugRenderer();
