+++ Voxengine/src/Editor/Panels/SceneHierarchyPanel.cpp
@@ -56,14 +56,16 @@ namespace Voxymore::Editor {
             }
         }
 
+#ifdef VXM_TERRAIN
     	if (s_Context->m_Terrain) s_Context->m_Terrain = std::make_unique<Terrain>();
 
 		ImGuiLib::DrawAssetField("Terrain Compute Shader", s_Context->m_Terrain->m_ComputeShaderField);
 		ImGuiLib::DrawAssetField("Terrain Grahics Shader", s_Context->m_Terrain->m_GraphicShaderField);
+#endif
 
         ImGui::Separator();
 
-        auto transformView = s_Context->m_Registry.view<TagComponent>();
+    	auto transformView = s_Context->m_Registry.view<TagComponent>();
         for (auto entity : transformView)
         {
             DrawEntityNode({entity, s_Context.get()});
